rename packages to fit with java standard
