+++ shared/src/test/java/passoff/TestFactory.java
@@ -1,4 +1,4 @@
-package passoffTests;
+package passoff;
 
 import chess.*;
 import org.junit.jupiter.api.Assertions;
