+++ shared/src/test/java/passoff/chess/piece/RookMoveTests.java
@@ -1,75 +1,72 @@
-package passoffTests.chessTests.chessPieceTests;
+package passoff.chess.piece;
 
 import org.junit.jupiter.api.Test;
 
-import static passoffTests.TestFactory.*;
+import static passoff.TestFactory.*;
 
-public class QueenMoveTests {
+public class RookMoveTests {
+
     @Test
-    public void queenMoveUntilEdge() {
+    public void rookMoveUntilEdge() {
+
         validateMoves("""
                         | | | | | | | | |
-                        | | | | | | |q| |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
+                        | | |R| | | | | |
                         | | | | | | | | |
                         """,
-                startPosition(7, 7),
+                startPosition(2, 3),
                 endPositions(new int[][]{
-                        {8, 7},
-                        {8, 8},
-                        {7, 8},
-                        {6, 8},
-                        {6, 7}, {5, 7}, {4, 7}, {3, 7}, {2, 7}, {1, 7},
-                        {6, 6}, {5, 5}, {4, 4}, {3, 3}, {2, 2}, {1, 1},
-                        {7, 6}, {7, 5}, {7, 4}, {7, 3}, {7, 2}, {7, 1},
-                        {8, 6},
+                        {2, 4}, {2, 5}, {2, 6}, {2, 7}, {2, 8},
+                        {2, 2}, {2, 1},
+                        {1, 3},
+                        {3, 3}, {4, 3}, {5, 3}, {6, 3}, {7, 3}, {8, 3},
                 })
         );
     }
 
 
     @Test
-    public void queenCaptureEnemy() {
+    public void rookCaptureEnemy() {
         validateMoves("""
-                        |b| | | | | | | |
                         | | | | | | | | |
-                        | | |R| | | | | |
                         | | | | | | | | |
-                        |Q| | |p| | | | |
                         | | | | | | | | |
-                        |P| |n| | | | | |
+                        |N| | | | | | | |
+                        |r| | | | |B| | |
+                        | | | | | | | | |
+                        |q| | | | | | | |
                         | | | | | | | | |
                         """,
                 startPosition(4, 1),
                 endPositions(new int[][]{
-                        {5, 1}, {6, 1}, {7, 1}, {8, 1},
-                        {5, 2},
-                        {4, 2}, {4, 3}, {4, 4},
-                        {3, 1}, {3, 2},
-                        {2, 3},
+                        {5, 1},
+                        {3, 1},
+                        {4, 2}, {4, 3}, {4, 4}, {4, 5}, {4, 6},
                 })
         );
     }
 
 
     @Test
-    public void queenBlocked() {
+    public void rookBlocked() {
         validateMoves("""
+                        | | | | | | |n|r|
+                        | | | | | | | |p|
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
                         | | | | | | | | |
-                        |P|R| | | | | | |
-                        |Q|K| | | | | | |
                         """,
-                startPosition(1, 1),
+                startPosition(8, 8),
                 endPositions(new int[][]{})
         );
     }
+
 }
